# news_swagger.json
{
    "openapi": "3.0.1",
    "info": {
    "title": "News API",
    "description": "A comprehensive News API with JWT authentication for posting and managing news articles with MongoDB",
    "contact": {
    "name": "News API Support",
    "email": "support@newsapi.com"
    },
    "version": "v1"
    },
    "paths": {
...
    "/api/News": {
    "get": {
    "tags": [
    "News"
    ],
    "summary": "Get all news articles",
    "responses": {
    "200": {
    "description": "Returns the list of news articles",
    "content": {
    "application/json": {
    "schema": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/News"
    }
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    },
    "post": {
    "tags": [
    "News"
    ],
    "summary": "Create a new news article (requires authentication)",
    "requestBody": {
    "description": "The news article data to create",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CreateNewsDto"
    }
    },
    "text/json": {
    "schema": {
    "$ref": "#/components/schemas/CreateNewsDto"
    }
    },
    "application/*+json": {
    "schema": {
    "$ref": "#/components/schemas/CreateNewsDto"
    }
    }
    }
    },
    "responses": {
    "201": {
    "description": "News article created successfully",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/News"
    }
    }
    }
    },
    "400": {
    "description": "Invalid input data",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "401": {
    "description": "Authentication required",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
    "/api/News/{id}": {
    "get": {
    "tags": [
    "News"
    ],
    "summary": "Get news by ID",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "description": "The MongoDB ObjectId of the news article",
    "required": true,
    "schema": {
    "type": "string"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Returns the news article",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/News"
    }
    }
    }
    },
    "404": {
    "description": "News article not found",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    },
    "put": {
    "tags": [
    "News"
    ],
    "summary": "Update an existing news article (requires authentication)",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "description": "The news article ID to update",
    "required": true,
    "schema": {
    "type": "string"
    }
    }
    ],
    "requestBody": {
    "description": "The news article data to update",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UpdateNewsDto"
    }
    },
    "text/json": {
    "schema": {
    "$ref": "#/components/schemas/UpdateNewsDto"
    }
    },
    "application/*+json": {
    "schema": {
    "$ref": "#/components/schemas/UpdateNewsDto"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "News article updated successfully",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/News"
    }
    }
    }
    },
    "400": {
    "description": "Invalid input data",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "401": {
    "description": "Authentication required",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "404": {
    "description": "News article not found",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    },
    "delete": {
    "tags": [
    "News"
    ],
    "summary": "Delete a news article (requires authentication)",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "description": "The news article ID to delete",
    "required": true,
    "schema": {
    "type": "string"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "News article deleted successfully"
    },
    "401": {
    "description": "Authentication required",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "404": {
    "description": "News article not found",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
...
    "/api/News/category/{category}": {
    "get": {
    "tags": [
    "News"
    ],
    "summary": "Get news by category",
    "parameters": [
    {
    "name": "category",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "application/json": {
    "schema": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/News"
    }
    }
    }
    }
    }
    }
    }
    },
    "/api/News/search": {
    "get": {
    "tags": [
    "News"
    ],
    "summary": "Search news articles",
    "parameters": [
    {
    "name": "searchTerm",
    "in": "query",
    "schema": {
    "type": "string"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "application/json": {
    "schema": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/News"
    }
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
...
    "CreateNewsDto": {
    "required": [
    "author",
    "content",
    "title"
    ],
    "type": "object",
    "properties": {
    "title": {
    "maxLength": 200,
    "minLength": 5,
    "type": "string"
    },
    "content": {
    "maxLength": 10000,
    "minLength": 10,
    "type": "string"
    },
    "author": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string"
    },
    "userId": {
    "maxLength": 24,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "category": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "tags": {
    "type": "array",
    "items": {
    "type": "string"
    },
    "nullable": true
    },
    "isPublished": {
    "type": "boolean"
    },
    "summary": {
    "maxLength": 500,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "imageUrl": {
    "type": "string",
    "format": "uri",
    "nullable": true
    },
    "publishedDate": {
    "type": "string",
    "format": "date-time",
    "nullable": true
    }
    },
    "additionalProperties": false
    },
...
    "News": {
    "type": "object",
    "properties": {
    "id": {
    "type": "string",
    "nullable": true
    },
    "title": {
    "type": "string",
    "nullable": true
    },
    "content": {
    "type": "string",
    "nullable": true
    },
    "author": {
    "type": "string",
    "nullable": true
    },
    "userId": {
    "type": "string",
    "nullable": true
    },
    "publishedDate": {
    "type": "string",
    "format": "date-time"
    },
    "category": {
    "type": "string",
    "nullable": true
    },
    "tags": {
    "type": "array",
    "items": {
    "type": "string"
    },
    "nullable": true
    },
    "isPublished": {
    "type": "boolean"
    },
    "summary": {
    "type": "string",
    "nullable": true
    },
    "imageUrl": {
    "type": "string",
    "nullable": true
    },
    "createdAt": {
    "type": "string",
    "format": "date-time"
    },
    "updatedAt": {
    "type": "string",
    "format": "date-time"
    }
    },
    "additionalProperties": false
    },
    "ProblemDetails": {
    "type": "object",
    "properties": {
    "type": {
    "type": "string",
    "nullable": true
    },
    "title": {
    "type": "string",
    "nullable": true
    },
    "status": {
    "type": "integer",
    "format": "int32",
    "nullable": true
    },
    "detail": {
    "type": "string",
    "nullable": true
    },
    "instance": {
    "type": "string",
    "nullable": true
    }
    },
    "additionalProperties": {}
    },
...
    "UpdateNewsDto": {
    "type": "object",
    "properties": {
    "title": {
    "maxLength": 200,
    "minLength": 5,
    "type": "string",
    "nullable": true
    },
    "content": {
    "maxLength": 10000,
    "minLength": 10,
    "type": "string",
    "nullable": true
    },
    "author": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "category": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "tags": {
    "type": "array",
    "items": {
    "type": "string"
    },
    "nullable": true
    },
    "isPublished": {
    "type": "boolean",
    "nullable": true
    },
    "summary": {
    "maxLength": 500,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "imageUrl": {
    "type": "string",
    "format": "uri",
    "nullable": true
    },
    "publishedDate": {
    "type": "string",
    "format": "date-time",
    "nullable": true
    }
    },
    "additionalProperties": false
    },
...
    },
    "securitySchemes": {
    "Bearer": {
    "type": "apiKey",
    "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
    "name": "Authorization",
    "in": "header"
    }
    }
    },
    "security": [
    {
    "Bearer": []
    }
    ]
    }